Update.platform.sh simplified, added error checking
Buddy: Jonty D

+++ update.platform.sh
@@ -1,64 +1,45 @@
 #!/bin/bash
 
-usage() { echo "Usage: $0 -p android|ios|osx [-c]"; echo "  -p -> platform, ios or android or osx (required)"; echo "  -c -> cpp11 support"; 1>&2; exit 1; }
-
 baseUrl="http://s3.amazonaws.com/eegeo-static/"
-srcPackageName="INVALID"
 destPackageName="./sdk.package.tar.gz"
-includeDestination="INVALID"
-sdkDestination="INVALID"
+srcPackageName="sdk.package.osx"
+includeDestination="./eegeo-platform"
+sdkDestination="sdk.package.osx"
 
-while getopts "p:c" o; do
-    case "${o}" in
-        p)
-            p=${OPTARG}
-            if [ "$p" != "ios" ]; then
-               if [ "$p" != "android" ]; then
-                  if [ "$p" != "osx" ]; then
-                    usage
-                  fi
-               fi
-            fi
-            ;;
-        c)
-            c="cpp11"
-            ;;
-        *)
-            usage
-            ;;
-    esac
-done
-shift $((OPTIND-1))
+echo "Updating osx platform..."
 
-if [ -z "${p}" ]; then
-    usage
-fi
-
-if [ "$p" == "ios" ]; then
-   srcPackageName="sdk.package.ios"
-   includeDestination="./ios/Include"
-   sdkDestination="sdk.package"
-elif [ "$p" == "android" ]; then
-   srcPackageName="sdk.package.android"
-   includeDestination="./android/libs"
-   sdkDestination="sdk.package.android"
-elif [ "$p" == "osx" ]; then
-   srcPackageName="sdk.package.osx"
-   includeDestination="./eegeo-platform"
-   sdkDestination="sdk.package.osx"   
-fi
-
-if [ "$c" == "cpp11" ]; then
-   srcPackageName="$srcPackageName.cpp11.tar.gz"
-else
-   srcPackageName="$srcPackageName.tar.gz"
-fi
-
-echo "Updating $p platform..."
 rm -rf $includeDestination
-curl $baseUrl$srcPackageName > ./$destPackageName
-tar -zxvf $destPackageName
+
+packageUrl="${baseUrl}${srcPackageName}.tar.gz"
+echo "downloading ${packageUrl} ..."
+curl $packageUrl > ./$destPackageName
+
+if [ $? -ne 0 ] ; then
+  echo "Failed to download SDK package from ${packageUrl}."
+  exit 1
+fi
+
+echo "unzipping ${destPackageName} ..."
+tar -zxvf $destPackageName 2> /dev/null
+
+if [ $? -ne 0 ] ; then
+  echo "Failed to unzip ${destPackageName}"
+  exit 1
+fi
+
 rm -f ./$destPackageName
+
 platformVersion=`cat ./$sdkDestination/version.txt`
+
 echo "Platform version --> $platformVersion"
+
 mv ./$sdkDestination $includeDestination
+
+if [ $? -ne 0 ] ; then
+  echo "Failed to move SDK package to destination directory"
+  exit 1
+fi
+
+echo "Successfully updated osx SDK package."
+
+exit 0
